Completed splitting of functions and Piece class into individual files.
