+++ cursatetra.py
@@ -8,512 +8,11 @@ import json
 from ct_draw import *
 from ct_label import *
 from ct_cell import *
+from Piece import *
 # SECTION: VERSION CHECK
 if sys.version_info[0] < 3:
 	print("This game requires Python 3. Please install it and/or run this file with it.")
 	exit()
-"""
-Class for active block data
-
-* Valid Constructor Values:
-	* y: [1, 19]
-		* Smallest block height-wise is Square, S, and Z; All take up two cells vertically
-	* x: [1, 10]
-	* pID: {'C', 'S', 'Z', 'L', 'R', 'I', 'T'}
-	* orient: {'', 'H', 'V', 'HP', 'VP'}
-		* Null-orientation is only valid for the Square
-
-* Methods & Valid Values:
-	* draw(): Draws the piece on the board
-	* undraw(): Erases the piece from the board
-	* getGhostDepth(): Determines the y-position of the ghost piece
-	* drawGhost(): Draws the ghost piece on the board
-	* undrawGhost(): Erases the ghost piece from the board
-	* getNewOrient(rotDir): Gets the new orientation for the piece based on rot. dir.
-		* rotDir: {'CW', 'CCW'}
-			* Clockwise or Counter-Clockwise
-			* Only needed for L, R, and T pieces
-	* rotate(rotDir): Rotates the piece in the indicated rot. dir.
-		* rotDir: See notes in getNewOrient
-	* move(direction): Moves the piece in the indicated direction
-		* direction: {'L', 'R', 'D'}
-			* Left, Right, or Down
-	* canRotate(rotDir): Returns True if the piece can rotate in the indicated rot. dir.
-		* rotDir: See notes in getNewOrient
-	* canMove(direction): Returns True if the piece can move in the indicated direction
-		* direction: See notes in move
-"""
-class Piece:
-	def __init__(self, y, x, p, o, board):
-		self.y = y
-		self.x = x
-		self.pID = p
-		self.orient = o
-		self.hasLanded = False
-		self.ghostDepth = 0
-		self.drawGhost(board)
-		self.draw()
-	def draw(self):
-		drawPiece(self.y, 2 * self.x - 1, self.orient, self.pID, wBoard, [crs.ACS_CKBOARD])
-	def undraw(self):
-		drawPiece(self.y, 2 * self.x - 1, self.orient, self.pID, wBoard, ". ")
-	def getGhostDepth(self, board):
-		originalY = self.y
-		while self.canMove('D', board):
-			self.y += 1
-		self.ghostDepth = self.y
-		self.y = originalY
-	def drawGhost(self, board):
-		if not doGhost:
-			return None
-		self.getGhostDepth(board)
-		drawPiece( \
-			self.ghostDepth, 2 * self.x - 1, \
-			self.orient, self.pID, wBoard, \
-			ghostChars \
-		)
-	def undrawGhost(self, board):
-		if not doGhost:
-			return None
-		drawPiece(self.ghostDepth, 2 * self.x - 1, self.orient, self.pID, board, ". ")
-	def getNewOrient(self, rotDir):
-		if self.pID == 'C':
-			return ''
-		elif self.pID in ('S', 'Z', 'I'):
-			return 'V' if self.orient == 'H' else 'H'
-		elif self.pID in ('L', 'R', 'T'):
-			if self.orient == 'H':
-				if rotDir == 'CW':
-					return 'V'
-				elif rotDir == 'CCW':
-					return 'VP'
-			elif self.orient == 'V':
-				if rotDir == 'CW':
-					return 'HP'
-				elif rotDir == 'CCW':
-					return 'H'
-			elif self.orient == 'HP':
-				if rotDir == 'CW':
-					return 'VP'
-				elif rotDir == 'CCW':
-					return 'V'
-			elif self.orient == 'VP':
-				if rotDir == 'CW':
-					return 'H'
-				elif rotDir == 'CCW':
-					return 'HP'
-	def rotate(self, rotDir, board):
-		self.undraw()
-		self.undrawGhost(board)
-		self.orient = self.getNewOrient(rotDir)
-		self.drawGhost(board)
-		self.draw()
-	def move(self, direction, board):
-		yShift = 1 if direction == 'D' else 0
-		xShift = 1 if direction == 'R' else -1 if direction == 'L' else 0
-		self.undraw()
-		self.undrawGhost(board)
-		self.y += yShift
-		self.x += xShift
-		self.drawGhost(board)
-		self.draw()
-	def canRotate(self, direction, board):
-		if self.pID == 'S':
-			if self.orient == 'H' and isCellEmpty(self.y, self.x, board) and \
-				isCellEmpty(self.y + 1, self.x, board):
-				return True
-			elif self.orient == 'V' and isCellEmpty(self.y + 1, self.x + 2, board) and \
-				isCellEmpty(self.y + 2, self.x, board) and \
-				isCellInBounds(self.y + 2, self.x + 2):
-				return True
-		elif self.pID == 'Z':
-			if self.orient == 'H' and isCellEmpty(self.y, self.x + 1, board) and \
-				isCellEmpty(self.y + 2, self.x, board):
-				return True
-			elif self.orient == 'V' and isCellEmpty(self.y + 2, self.x + 1, board) and \
-				isCellEmpty(self.y + 2, self.x + 2, board) and \
-				isCellInBounds(self.y + 2, self.x + 2):
-				return True
-		elif self.pID == 'L':
-			if self.orient == 'H':
-				if direction == 'CW' and isCellInBounds(self.y + 2, self.x) and \
-					isCellEmpty(self.y + 1, self.x, board) and isCellEmpty(self.y + 2, self.x, board) and \
-					isCellEmpty(self.y + 1, self.x + 2, board):
-					return True
-				elif direction == 'CCW' and isCellInBounds(self.y + 2, self.x) and \
-					isCellEmpty(self.y + 1, self.x, board) and isCellEmpty(self.y, self.x + 2, board) and \
-					isCellEmpty(self.y + 1, self.x + 2, board):
-					return True
-			elif self.orient == 'V':
-				if direction == 'CW' and isCellEmpty(self.y, self.x, board) and \
-					isCellEmpty(self.y, self.x + 1, board) and \
-					isCellEmpty(self.y + 2, self.x + 1, board):
-					return True
-				elif direction == 'CCW' and isCellEmpty(self.y, self.x + 1, board) and \
-					isCellEmpty(self.y + 2, self.x + 1, board) and \
-					isCellEmpty(self.y + 2, self.x + 2, board):
-					return True
-			elif self.orient == 'HP':
-				if direction == 'CW' and isCellInBounds(self.y + 2, self.x + 2) and \
-					isCellEmpty(self.y + 1, self.x, board) and isCellEmpty(self.y, self.x + 2, board) and \
-					isCellEmpty(self.y + 1, self.x + 2, board):
-					return True
-				elif direction == 'CCW' and isCellInBounds(self.y + 2, self.x + 2) and \
-					isCellEmpty(self.y + 1, self.x, board) and isCellEmpty(self.y + 2, self.x, board) and \
-					isCellEmpty(self.y + 1, self.x + 2, board):
-					return True
-			elif self.orient == 'VP':
-				if direction == 'CW' and isCellEmpty(self.y, self.x + 1, board) and \
-					isCellEmpty(self.y + 2, self.x + 1, board) and \
-					isCellEmpty(self.y + 2, self.x + 2, board):
-					return True
-				elif direction == 'CCW' and isCellEmpty(self.y, self.x, board) and \
-					isCellEmpty(self.y, self.x + 1, board) and \
-					isCellEmpty(self.y + 2, self.x + 1, board):
-					return True
-		elif self.pID == 'R':
-			if self.orient == 'H':
-				if direction == 'CW' and isCellInBounds(self.y + 2, self.x + 2) and \
-					isCellEmpty(self.y + 1, self.x, board) and isCellEmpty(self.y, self.x, board) and \
-					isCellEmpty(self.y + 1, self.x + 2, board):
-					return True
-				elif direction == 'CCW' and isCellInBounds(self.y + 2, self.x + 2) and \
-					isCellEmpty(self.y + 1, self.x, board) and isCellEmpty(self.y + 2, self.x + 2, board) and \
-					isCellEmpty(self.y + 1, self.x + 2, board):
-					return True
-			elif self.orient == 'V':
-				if direction == 'CW' and isCellEmpty(self.y, self.x + 1, board) and \
-					isCellEmpty(self.y, self.x + 2, board) and \
-					isCellEmpty(self.y + 2, self.x + 1, board):
-					return True
-				elif direction == 'CCW' and isCellEmpty(self.y, self.x + 1, board) and \
-					isCellEmpty(self.y + 2, self.x + 1, board) and \
-					isCellEmpty(self.y + 2, self.x, board):
-					return True
-			elif self.orient == 'HP':
-				if direction == 'CW' and isCellInBounds(self.y + 2, self.x) and \
-					isCellEmpty(self.y + 1, self.x, board) and isCellEmpty(self.y + 2, self.x + 2, board) and \
-					isCellEmpty(self.y + 1, self.x + 2, board):
-					return True
-				elif direction == 'CCW' and isCellInBounds(self.y + 2, self.x) and \
-					isCellEmpty(self.y, self.x, board) and isCellEmpty(self.y + 1, self.x, board) and \
-					isCellEmpty(self.y + 1, self.x + 2, board):
-					return True
-			elif self.orient == 'VP':
-				if direction == 'CW' and isCellEmpty(self.y, self.x + 1, board) and \
-					isCellEmpty(self.y + 2, self.x, board) and \
-					isCellEmpty(self.y + 2, self.x + 1, board):
-					return True
-				elif direction == 'CCW' and isCellEmpty(self.y, self.x + 1, board) and \
-					isCellEmpty(self.y, self.x + 2, board) and \
-					isCellEmpty(self.y + 2, self.x + 1, board):
-					return True
-		elif self.pID == 'I':
-			if self.orient == 'H' and isCellInBounds(self.y + 3, self.x + 3) and \
-				isCellInBounds(self.y + 3, self.x) and \
-				isCellEmpty(self.y + 2, self.x, board) and \
-				isCellEmpty(self.y + 2, self.x + 2, board) and \
-				isCellEmpty(self.y + 2, self.x + 3, board):
-				return True
-			elif self.orient == 'V' and isCellInBounds(self.y + 3, self.x + 1) and \
-				isCellEmpty(self.y, self.x + 1, board) and \
-				isCellEmpty(self.y + 1, self.x + 1, board) and \
-				isCellEmpty(self.y + 3, self.x + 1, board):
-				return True
-		elif self.pID == 'T':
-			if self.orient == 'H' and isCellEmpty(self.y, self.x + 1, board):
-				return True
-			elif self.orient == 'V' and isCellEmpty(self.y + 1, self.x + 2, board) and \
-				isCellInBounds(self.y + 2, self.x + 2):
-				return True
-			elif self.orient == 'HP' and isCellEmpty(self.y + 2, self.x + 1, board):
-				return True
-			elif self.orient == 'VP' and isCellEmpty(self.y + 1, self.x, board) and \
-				isCellInBounds(self.y + 2, self.x):
-				return True
-		return False
-	def canMove(self, direction, board):
-		if self.pID == 'C':
-			if direction == 'L' and isCellInBounds(self.y + 1, self.x - 1) and \
-				isCellEmpty(self.y, self.x - 1, board) and \
-				isCellEmpty(self.y + 1, self.x - 1, board):
-				return True
-			elif direction == 'R' and isCellInBounds(self.y + 1, self.x + 2) and \
-				isCellEmpty(self.y, self.x + 2, board) and \
-				isCellEmpty(self.y + 1, self.x + 2, board):
-				return True
-			elif direction == 'D' and isCellInBounds(self.y + 2, self.x) and \
-				isCellEmpty(self.y + 2, self.x, board) and \
-				isCellEmpty(self.y + 2, self.x + 1, board):
-				return True
-		elif self.pID == 'S':
-			if self.orient == 'H':
-				if direction == 'L' and isCellInBounds(self.y + 2, self.x - 1) and \
-					isCellEmpty(self.y + 2, self.x - 1, board) and \
-					isCellEmpty(self.y + 1, self.x, board):
-					return True
-				elif direction == 'R' and isCellInBounds(self.y + 2, self.x + 3) and \
-					isCellEmpty(self.y + 1, self.x + 3, board) and \
-					isCellEmpty(self.y + 2, self.x + 2, board):
-					return True
-				elif direction == 'D' and isCellInBounds(self.y + 3, self.x) and \
-					isCellEmpty(self.y + 3, self.x, board) and \
-					isCellEmpty(self.y + 3, self.x + 1, board) and \
-					isCellEmpty(self.y + 2, self.x + 2, board):
-					return True
-			elif self.orient == 'V':
-				if direction == 'L' and isCellInBounds(self.y + 2, self.x - 1) and \
-					isCellEmpty(self.y, self.x - 1, board) and \
-					isCellEmpty(self.y + 1, self.x - 1, board) and \
-					isCellEmpty(self.y + 2, self.x, board):
-					return True
-				elif direction == 'R' and isCellInBounds(self.y + 2, self.x + 2) and \
-					isCellEmpty(self.y + 1, self.x + 2, board) and \
-					isCellEmpty(self.y + 2, self.x + 2, board) and \
-					isCellEmpty(self.y, self.x + 1, board):
-					return True
-				elif direction == 'D' and isCellInBounds(self.y + 3, self.x + 1) and \
-					isCellEmpty(self.y + 3, self.x + 1, board) and \
-					isCellEmpty(self.y + 2, self.x, board):
-					return True
-		elif self.pID == 'Z':
-			if self.orient == 'H':
-				if direction == 'L' and isCellInBounds(self.y + 2, self.x - 1) and \
-					isCellEmpty(self.y + 1, self.x - 1, board) and \
-					isCellEmpty(self.y + 2, self.x, board):
-					return True
-				elif direction == 'R' and isCellInBounds(self.y + 2, self.x + 3) and \
-					isCellEmpty(self.y + 2, self.x + 3, board) and \
-					isCellEmpty(self.y + 1, self.x + 2, board):
-					return True
-				elif direction == 'D' and isCellInBounds(self.y + 3, self.x) and \
-					isCellEmpty(self.y + 3, self.x + 1, board) and \
-					isCellEmpty(self.y + 3, self.x + 2, board) and \
-					isCellEmpty(self.y + 2, self.x, board):
-					return True
-			elif self.orient == 'V':
-				if direction == 'L' and isCellInBounds(self.y + 2, self.x - 1) and \
-					isCellEmpty(self.y + 1, self.x - 1, board) and \
-					isCellEmpty(self.y + 2, self.x - 1, board) and \
-					isCellEmpty(self.y, self.x, board):
-					return True
-				elif direction == 'R' and isCellInBounds(self.y + 2, self.x + 2) and \
-					isCellEmpty(self.y, self.x + 2, board) and \
-					isCellEmpty(self.y + 1, self.x + 2, board) and \
-					isCellEmpty(self.y + 2, self.x + 1, board):
-					return True
-				elif direction == 'D' and isCellInBounds(self.y + 3, self.x) and \
-					isCellEmpty(self.y + 3, self.x, board) and \
-					isCellEmpty(self.y + 2, self.x + 1, board):
-					return True
-		elif self.pID == 'L':
-			if self.orient == 'H':
-				if direction == 'L' and isCellInBounds(self.y + 2, self.x) and \
-					isCellEmpty(self.y, self.x, board) and \
-					isCellEmpty(self.y + 1, self.x, board) and \
-					isCellEmpty(self.y + 2, self.x, board):
-					return True
-				elif direction == 'R' and isCellInBounds(self.y + 2, self.x + 3) and \
-					isCellEmpty(self.y + 2, self.x + 3, board) and \
-					isCellEmpty(self.y + 1, self.x + 2, board) and \
-					isCellEmpty(self.y, self.x + 2, board):
-					return True
-				elif direction == 'D' and isCellInBounds(self.y + 3, self.x + 1) and \
-					isCellEmpty(self.y + 3, self.x + 1, board) and \
-					isCellEmpty(self.y + 3, self.x + 2, board):
-					return True
-			elif self.orient == 'V':
-				if direction == 'L' and isCellInBounds(self.y + 2, self.x - 1) and \
-					isCellEmpty(self.y + 1, self.x - 1, board) and \
-					isCellEmpty(self.y + 2, self.x - 1, board):
-					return True
-				elif direction == 'R' and isCellInBounds(self.y + 2, self.x + 3) and \
-					isCellEmpty(self.y + 1, self.x + 3, board) and \
-					isCellEmpty(self.y + 2, self.x + 1, board):
-					return True
-				elif direction == 'D' and isCellInBounds(self.y + 3, self.x) and \
-					isCellEmpty(self.y + 3, self.x, board) and \
-					isCellEmpty(self.y + 2, self.x + 1, board) and \
-					isCellEmpty(self.y + 2, self.x + 2, board):
-					return True
-			if self.orient == 'HP':
-				if direction == 'L' and isCellInBounds(self.y + 2, self.x - 1) and \
-					isCellEmpty(self.y, self.x - 1, board) and \
-					isCellEmpty(self.y + 1, self.x, board) and \
-					isCellEmpty(self.y + 2, self.x, board):
-					return True
-				elif direction == 'R' and isCellInBounds(self.y + 2, self.x + 2) and \
-					isCellEmpty(self.y, self.x + 2, board) and \
-					isCellEmpty(self.y + 1, self.x + 2, board) and \
-					isCellEmpty(self.y + 2, self.x + 2, board):
-					return True
-				elif direction == 'D' and isCellInBounds(self.y + 3, self.x + 1) and \
-					isCellEmpty(self.y + 3, self.x + 1, board) and \
-					isCellEmpty(self.y + 1, self.x, board):
-					return True
-			elif self.orient == 'VP':
-				if direction == 'L' and isCellInBounds(self.y + 1, self.x - 1) and \
-					isCellEmpty(self.y + 1, self.x - 1, board) and \
-					isCellEmpty(self.y, self.x + 1, board):
-					return True
-				elif direction == 'R' and isCellInBounds(self.y + 1, self.x + 3) and \
-					isCellEmpty(self.y, self.x + 3, board) and \
-					isCellEmpty(self.y + 1, self.x + 3, board):
-					return True
-				elif direction == 'D' and isCellInBounds(self.y + 2, self.x) and \
-					isCellEmpty(self.y + 2, self.x, board) and \
-					isCellEmpty(self.y + 2, self.x + 1, board) and \
-					isCellEmpty(self.y + 2, self.x + 2, board):
-					return True
-		elif self.pID == 'R':
-			if self.orient == 'H':
-				if direction == 'L' and isCellInBounds(self.y + 2, self.x - 1) and \
-					isCellEmpty(self.y + 2, self.x - 1, board) and \
-					isCellEmpty(self.y + 1, self.x, board) and \
-					isCellEmpty(self.y, self.x, board):
-					return True
-				elif direction == 'R' and isCellInBounds(self.y + 2, self.x + 2) and \
-					isCellEmpty(self.y, self.x + 2, board) and \
-					isCellEmpty(self.y + 1, self.x + 2, board) and \
-					isCellEmpty(self.y + 2, self.x + 2, board):
-					return True
-				elif direction == 'D' and isCellInBounds(self.y + 3, self.x) and \
-					isCellEmpty(self.y + 3, self.x, board) and \
-					isCellEmpty(self.y + 3, self.x + 1, board):
-					return True
-			elif self.orient == 'V':
-				if direction == 'L' and isCellInBounds(self.y + 1, self.x - 1) and \
-					isCellEmpty(self.y, self.x - 1, board) and \
-					isCellEmpty(self.y + 1, self.x - 1, board):
-					return True
-				elif direction == 'R' and isCellInBounds(self.y + 1, self.x + 3) and \
-					isCellEmpty(self.y + 1, self.x + 3, board) and \
-					isCellEmpty(self.y, self.x + 1, board):
-					return True
-				elif direction == 'D' and isCellInBounds(self.y + 2, self.x) and \
-					isCellEmpty(self.y + 2, self.x, board) and \
-					isCellEmpty(self.y + 2, self.x + 1, board) and \
-					isCellEmpty(self.y + 2, self.x + 2, board):
-					return True
-			if self.orient == 'HP':
-				if direction == 'L' and isCellInBounds(self.y + 2, self.x) and \
-					isCellEmpty(self.y, self.x, board) and \
-					isCellEmpty(self.y + 1, self.x, board) and \
-					isCellEmpty(self.y + 2, self.x, board):
-					return True
-				elif direction == 'R' and isCellInBounds(self.y, self.x + 3) and \
-					isCellEmpty(self.y, self.x + 3, board) and \
-					isCellEmpty(self.y + 1, self.x + 2, board) and \
-					isCellEmpty(self.y + 2, self.x + 2, board):
-					return True
-				elif direction == 'D' and isCellInBounds(self.y + 3, self.x + 1) and \
-					isCellEmpty(self.y + 3, self.x + 1, board) and \
-					isCellEmpty(self.y + 1, self.x + 2, board):
-					return True
-			elif self.orient == 'VP':
-				if direction == 'L' and isCellInBounds(self.y + 2, self.x - 1) and \
-					isCellEmpty(self.y + 1, self.x - 1, board) and \
-					isCellEmpty(self.y + 2, self.x + 1, board):
-					return True
-				elif direction == 'R' and isCellInBounds(self.y + 2, self.x + 3) and \
-					isCellEmpty(self.y + 1, self.x + 3, board) and \
-					isCellEmpty(self.y + 2, self.x + 3, board):
-					return True
-				elif direction == 'D' and isCellInBounds(self.y + 3, self.x + 2) and \
-					isCellEmpty(self.y + 3, self.x + 2, board) and \
-					isCellEmpty(self.y + 2, self.x + 1, board) and \
-					isCellEmpty(self.y + 2, self.x, board):
-					return True
-		elif self.pID == 'I':
-			if self.orient == 'H':
-				if direction == 'L' and isCellInBounds(self.y + 3, self.x) and \
-					isCellEmpty(self.y, self.x, board) and \
-					isCellEmpty(self.y + 1, self.x, board) and \
-					isCellEmpty(self.y + 2, self.x, board) and \
-					isCellEmpty(self.y + 3, self.x, board):
-					return True
-				elif direction == 'R' and isCellInBounds(self.y + 3, self.x + 2) and \
-					isCellEmpty(self.y, self.x + 2, board) and \
-					isCellEmpty(self.y + 1, self.x + 2, board) and \
-					isCellEmpty(self.y + 2, self.x + 2, board) and \
-					isCellEmpty(self.y + 3, self.x + 2, board):
-					return True
-				elif direction == 'D' and isCellInBounds(self.y + 4, self.x + 1) and \
-					isCellEmpty(self.y + 4, self.x + 1, board):
-					return True
-			elif self.orient == 'V':
-				if direction == 'L' and isCellInBounds(self.y + 2, self.x - 1) and \
-					isCellEmpty(self.y + 2, self.x - 1, board):
-					return True
-				elif direction == 'R' and isCellInBounds(self.y + 2, self.x + 4) and \
-					isCellEmpty(self.y + 2, self.x + 4, board):
-					return True
-				elif direction == 'D' and isCellInBounds(self.y + 3, self.x) and \
-					isCellEmpty(self.y + 3, self.x, board) and \
-					isCellEmpty(self.y + 3, self.x + 1, board) and \
-					isCellEmpty(self.y + 3, self.x + 2, board) and \
-					isCellEmpty(self.y + 3, self.x + 3, board):
-					return True
-		elif self.pID == 'T':
-			if self.orient == 'H':
-				if direction == 'L' and isCellInBounds(self.y + 2, self.x - 1) and \
-					isCellEmpty(self.y + 1, self.x - 1, board) and \
-					isCellEmpty(self.y + 2, self.x, board):
-					return True
-				elif direction == 'R' and isCellInBounds(self.y + 2, self.x + 3) and \
-					isCellEmpty(self.y + 1, self.x + 3, board) and \
-					isCellEmpty(self.y + 2, self.x + 2, board):
-					return True
-				elif direction == 'D' and isCellInBounds(self.y + 3, self.x) and \
-					isCellEmpty(self.y + 3, self.x + 1, board) and \
-					isCellEmpty(self.y + 2, self.x + 2, board) and \
-					isCellEmpty(self.y + 2, self.x, board):
-					return True
-			elif self.orient == 'V':
-				if direction == 'L' and isCellInBounds(self.y + 2, self.x - 1) and \
-					isCellEmpty(self.y + 1, self.x - 1, board) and \
-					isCellEmpty(self.y + 2, self.x, board) and \
-					isCellEmpty(self.y, self.x, board):
-					return True
-				elif direction == 'R' and isCellInBounds(self.y + 2, self.x + 2) and \
-					isCellEmpty(self.y, self.x + 2, board) and \
-					isCellEmpty(self.y + 1, self.x + 2, board) and \
-					isCellEmpty(self.y + 2, self.x + 2, board):
-					return True
-				elif direction == 'D' and isCellInBounds(self.y + 3, self.x) and \
-					isCellEmpty(self.y + 3, self.x + 1, board) and \
-					isCellEmpty(self.y + 2, self.x, board):
-					return True
-			elif self.orient == 'HP':
-				if direction == 'L' and isCellInBounds(self.y + 1, self.x - 1) and \
-					isCellEmpty(self.y + 1, self.x - 1, board) and \
-					isCellEmpty(self.y, self.x, board):
-					return True
-				elif direction == 'R' and isCellInBounds(self.y + 1, self.x + 3) and \
-					isCellEmpty(self.y + 1, self.x + 3, board) and \
-					isCellEmpty(self.y, self.x + 2, board):
-					return True
-				elif direction == 'D' and isCellInBounds(self.y + 2, self.x) and \
-					isCellEmpty(self.y + 2, self.x, board) and \
-					isCellEmpty(self.y + 2, self.x + 1, board) and \
-					isCellEmpty(self.y + 2, self.x + 2, board):
-					return True
-			elif self.orient == 'VP':
-				if direction == 'L' and isCellInBounds(self.y + 2, self.x) and \
-					isCellEmpty(self.y, self.x, board) and \
-					isCellEmpty(self.y + 1, self.x, board) and \
-					isCellEmpty(self.y + 2, self.x, board):
-					return True
-				elif direction == 'R' and isCellInBounds(self.y + 2, self.x + 3) and \
-					isCellEmpty(self.y + 1, self.x + 3, board) and \
-					isCellEmpty(self.y + 2, self.x + 2, board) and \
-					isCellEmpty(self.y, self.x + 2, board):
-					return True
-				elif direction == 'D' and isCellInBounds(self.y + 3, self.x + 1) and \
-					isCellEmpty(self.y + 3, self.x + 1, board) and \
-					isCellEmpty(self.y + 2, self.x + 2, board):
-					return True
-		return False
 """
 Main function
 
@@ -834,10 +333,10 @@ def ctMain():
 			elif keypress == "ENTER":
 				continue
 			elif keypress == 'G':
-				if doGhost:
+				if doGhost[0]:
 					piece.undrawGhost(wBoard)
-				doGhost = not doGhost
-				if doGhost:
+				doGhost[0] = not doGhost[0]
+				if doGhost[0]:
 					piece.drawGhost(wBoard)
 				continue
 			# If neither the enter nor q key are pressed, it must be ESC,
